offers/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
django-cors-headers==3.8.0